Copy-paste text Bachelorthesis and literature list
The name of the file on how to write in R has been changed to
filterbubblesDENL_example.Rmd.
In the filterbubblesDENL.Rmd file, the introduction and related work
part of Lisanne's Bachelorthesis have been pasted.
Also the contents of the .bib file used for Lisanne's Bachelorthesis has
been pasted into the references.bib.